requirements.txt
pandas
streamlit
Pillow
bokeh==2.4.3
plotly
pybase64
nltk
matplotlib
seaborn
